move SqGyroPosIntGuideStep and SqPosIntDriveToStep from  SkystoneAutoBlue1 sample to AutoLib so it can be more widely shared. for now, leave the Vuforia-enhanced version (now called VfSqGyroPosIntGuideStep) in SkystoneAutoBlue1 sample, where it is used in that sample's local version of SqPosIntDriveToStep. reorder some classes in AutoLib to group like functionality together better.
