+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/_Test/_Drive/SkystoneAutoBlue1.java
@@ -27,122 +27,51 @@ import java.util.ArrayList;
 //@Disabled
 public class SkystoneAutoBlue1 extends OpMode {
 
-    // return done when we're within tolerance distance of target position
-    class PositionTerminatorStep extends AutoLib.MotorGuideStep {
-
-        OpMode mOpMode;
-        SensorLib.PositionIntegrator mPosInt;
-        Position mTarget;
-        double mTol;
-        double mPrevDist;
-
-        public PositionTerminatorStep(OpMode opmode, SensorLib.PositionIntegrator posInt, Position target, double tol) {
-            mOpMode = opmode;
-            mPosInt = posInt;
-            mTarget = target;
-            mTol = tol;
-            mPrevDist = 1e6;    // infinity
-        }
-
-        @Override
-        public boolean loop() {
-            super.loop();
-            Position current = mPosInt.getPosition();
-            double dist = Math.sqrt((mTarget.x-current.x)*(mTarget.x-current.x) + (mTarget.y-current.y)*(mTarget.y-current.y));
-            if (mOpMode != null) {
-                mOpMode.telemetry.addData("PTS target", String.format("%.2f", mTarget.x) + ", " + String.format("%.2f", mTarget.y));
-                mOpMode.telemetry.addData("PTS current", String.format("%.2f", current.x) + ", " + String.format("%.2f", current.y));
-                mOpMode.telemetry.addData("PTS dist", String.format("%.2f", dist));
-            }
-            boolean bDone = (dist < mTol);
-
-            // try to deal with "circling the drain" problem -- when we're close to the tolerance
-            // circle, but we can't turn sharply enough to get into it, we circle endlessly --
-            // if we detect that situation, just give up and move on.
-            // simple test: if we're close but the distance to the target increases, we've missed it.
-            if (dist < mTol*4 && dist > mPrevDist)
-                bDone = true;
-            mPrevDist = dist;
-
-            return bDone;
-        }
-    }
-
     // guide step that uses a gyro and a position integrator to determine how to guide the robot to the target
-    class SqGyroPosIntGuideStep extends AutoLib.SquirrelyGyroGuideStep {
+    class VfSqGyroPosIntGuideStep extends AutoLib.SqGyroPosIntGuideStep {
 
-        OpMode mOpMode;
-        Position mTarget;
-        SensorLib.EncoderGyroPosInt mPosInt;
-        double mTol;
-        float mMaxPower;
-        float mMinPower = 0.25f;
-        float mSgnPower;
+        VuforiaLib_SkyStone mVLib = null;       // (optional) Vuforia data source
+        SensorLib.EncoderGyroPosInt mPosInt;    // Encoder/gyro-based position integrator to keep track of where we are
 
-        public SqGyroPosIntGuideStep(OpMode opmode, SensorLib.EncoderGyroPosInt posInt, Position target, float heading,
-                                     SensorLib.PID pid, ArrayList<AutoLib.SetPower> motorsteps, float power, double tol) {
-            // this.preAdd(new SquirrelyGyroGuideStep(mode, direction, heading, gyro, pid, steps, power));
-            super(opmode, 0, heading, posInt.getGyro(), pid, motorsteps, power);
-            mTarget = target;
+        public VfSqGyroPosIntGuideStep(OpMode opmode, VuforiaLib_SkyStone vfLib, SensorLib.EncoderGyroPosInt posInt, Position target, float heading,
+                                     SensorLib.PID pid, ArrayList<AutoLib.SetPower> motorsteps, float power, double tol)
+        {
+            super(opmode, posInt, target, heading, pid, motorsteps, power, tol);
+            mVLib = vfLib;
             mPosInt = posInt;
-            mTol = tol;
-            mMaxPower = Math.abs(power);
-            mSgnPower = (power > 0) ? 1 : -1;
         }
 
         public boolean loop() {
-            // run the EncoderGyroPosInt to update its position based on encoders and gyro
-            mPosInt.loop();
 
-            // update Vuforia info and, if we have valid location data, update the position integrator with it
-            mVLib.loop(true);       // update Vuforia location info
+            // if we have Vuforia enabled, try to use info from it, too
+            if (mVLib != null) {
 
-            // don't believe Vuforia data if we're currently turning (blurry image?)
-            float angVel = mPosInt.getGyro().getHeadingVelocity();	// in deg/sec
-            boolean turningTooFast = Math.abs(angVel) > 10.0;
+                // update Vuforia info and, if we have valid location data, update the position integrator with it
+                mVLib.loop(true);       // update Vuforia location info
 
-            // if we have Vuforia location data, update the position integrator from it.
-            // use STATUS and STATUS_INFO associated with the sample to decide how much to believe it.
-            if (mVLib.haveLocation() && mVLib.getTrackableStatusInfo() == TrackableResult.STATUS_INFO.NORMAL && !turningTooFast) {
-                if (mVLib.getTrackableStatus() == TrackableResult.STATUS.TRACKED)
-                    mPosInt.setPosition(mVLib.getFieldPosition());
-                else
-                if (mVLib.getTrackableStatus() == TrackableResult.STATUS.EXTENDED_TRACKED)
-                    mPosInt.setPosition(mVLib.getFieldPosition(), 0.5f);
+                // don't believe Vuforia data if we're currently turning (blurry image?)
+                float angVel = mPosInt.getGyro().getHeadingVelocity();    // in deg/sec
+                boolean turningTooFast = Math.abs(angVel) > 10.0;
+
+                // if we have Vuforia location data, update the position integrator from it.
+                // use STATUS and STATUS_INFO associated with the sample to decide how much to believe it.
+                if (mVLib.haveLocation() && mVLib.getTrackableStatusInfo() == TrackableResult.STATUS_INFO.NORMAL && !turningTooFast) {
+                    if (mVLib.getTrackableStatus() == TrackableResult.STATUS.TRACKED)
+                        mPosInt.setPosition(mVLib.getFieldPosition());
+                    else if (mVLib.getTrackableStatus() == TrackableResult.STATUS.EXTENDED_TRACKED)
+                        mPosInt.setPosition(mVLib.getFieldPosition(), 0.5f);
+                }
             }
 
-            // update the SquirrelyGyroGuideStep direction to continue heading for the target
-            // while maintaining the heading (orientation) given for this Step
-            float direction = (float) HeadingToTarget(mTarget, mPosInt.getPosition());
-            super.setDirection(direction);
-
-            // when we're close to the target, reduce speed so we don't overshoot
-            Position current = mPosInt.getPosition();
-            float dist = (float)Math.sqrt((mTarget.x-current.x)*(mTarget.x-current.x) + (mTarget.y-current.y)*(mTarget.y-current.y));
-            float brakeDist = (float)mTol * 5;  // empirical ...
-            if (dist < brakeDist) {
-                float power = mSgnPower * (mMinPower + (mMaxPower-mMinPower)*(dist/brakeDist));
-                super.setMaxPower(power);
-            }
-
-            // run the underlying GyroGuideStep and return what it returns for "done" -
-            // currently, it leaves it up to the terminating step to end the Step
+            // run the base Step and return what it returns for "done" -
+            // note that we're running this code AFTER updating the posInt with Vuforia data ...
             return super.loop();
         }
 
-        private double HeadingToTarget(Position target, Position current) {
-            double headingXrad = Math.atan2((target.y - current.y), (target.x - current.x));        // pos CCW from X-axis
-            double headingYdeg = SensorLib.Utils.wrapAngle(Math.toDegrees(headingXrad) - 90.0);     // pos CCW from Y-axis
-            if (mOpMode != null) {
-                mOpMode.telemetry.addData("GPIGS.HTT target", String.format("%.2f", target.x) + ", " + String.format("%.2f", target.y));
-                mOpMode.telemetry.addData("GPIGS.HTT current", String.format("%.2f", current.x) + ", " + String.format("%.2f", current.y));
-                mOpMode.telemetry.addData("GPIGS.HTT heading", String.format("%.2f", headingYdeg));
-            }
-            return headingYdeg;
-        }
     }
 
     // Step: drive to given absolute field position while facing in the given direction using given EncoderGyroPosInt
+    // and (optional) Vuforia lib that supplies occasional position updates to that PositionIntegrator
     class SqPosIntDriveToStep extends AutoLib.GuidedTerminatedDriveStep {
 
         SensorLib.EncoderGyroPosInt mPosInt;
@@ -151,8 +80,8 @@ public class SkystoneAutoBlue1 extends OpMode {
         public SqPosIntDriveToStep(OpMode opmode, SensorLib.EncoderGyroPosInt posInt, DcMotor[] motors,
                                    float power, SensorLib.PID pid, Position target, float heading, double tolerance, boolean stop)
         {
-            super(opmode, new SqGyroPosIntGuideStep(opmode, posInt, target, heading, pid, null, power, tolerance),
-                    new PositionTerminatorStep(opmode, posInt, target, tolerance),
+            super(opmode, new VfSqGyroPosIntGuideStep(opmode, mVLib, posInt, target, heading, pid, null, power, tolerance),
+                    new AutoLib.PositionTerminatorStep(opmode, posInt, target, tolerance),
                     motors);
 
             mPosInt = posInt;
@@ -170,7 +99,7 @@ public class SkystoneAutoBlue1 extends OpMode {
     SensorLib.EncoderGyroPosInt mPosInt;    // Encoder/gyro-based position integrator to keep track of where we are
     SensorLib.PIDAdjuster mPidAdjuster;     // for interactive adjustment of PID parameters
     RobotHardware rh;                       // standard hardware set for these tests
-    VuforiaLib_SkyStone mVLib;
+    VuforiaLib_SkyStone mVLib = null;       // (optional) Vuforia data source
 
     /**
      * Constructor
@@ -195,7 +124,7 @@ public class SkystoneAutoBlue1 extends OpMode {
         telemetry.addData("", "autonomous point to point");
         telemetry.addData("", "navigation using PositionIntegrator");
         telemetry.addData("", "driven by motor encoders with ");
-        telemetry.addData("", "occasional Vuforia updates ");
+        telemetry.addData("", "optional occasional Vuforia updates ");
 
         // create a PID controller for the sequence
         // parameters of the PID controller for this sequence - assumes 20-gear motors (fast)
@@ -219,9 +148,12 @@ public class SkystoneAutoBlue1 extends OpMode {
                         SensorLib.EncoderGyroPosInt.DriveType.MECANUM;
         mPosInt = new SensorLib.EncoderGyroPosInt(dt,this, rh.mIMU, rh.mMotors, countsPerRev, wheelDiam, initialPosn);
 
-        // Start up Vuforia
-        mVLib = new VuforiaLib_SkyStone();
-        mVLib.init(this);     // pass it this OpMode (so it can do telemetry output)
+        // (option) Start up Vuforia
+        final boolean bUseVuforia = false;
+        if (bUseVuforia) {
+            mVLib = new VuforiaLib_SkyStone();
+            mVLib.init(this);     // pass it this OpMode (so it can do telemetry output)
+        }
 
         // create an autonomous sequence with the steps to drive
         // several legs of a polygonal course ---
@@ -285,7 +217,8 @@ public class SkystoneAutoBlue1 extends OpMode {
     @Override public void start()
     {
         /** Start tracking the data sets we care about. */
-        mVLib.start();
+        if (mVLib != null)
+            mVLib.start();
     }
 
     public void loop() {
@@ -309,7 +242,8 @@ public class SkystoneAutoBlue1 extends OpMode {
 
     public void stop() {
         super.stop();
-        mVLib.stop();
+        if (mVLib != null)
+            mVLib.stop();
     }
 }
